refactor - remove unused imports and comments
